requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==2.3.7
ccxt==4.1.64
APScheduler==3.10.4
python-telegram-bot==20.7
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
numpy==1.24.3 